WIP: move checkEnabled() to /lib
